pyproject.toml
[tool.poetry]
name = "openai-agent"
version = "0.1.0"
description = "An example openai agent"
authors = ["IBM"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">= 3.13, < 4.0"
python-dotenv = "^1.0.1"
jsonschema = "^4.23.0"
openai-agents = "^0.0.10"
uvenv = "^3.7.5"
uvx = "^1.0.0, <2.0"
nest-asyncio = "^1.6.0"
logfire = "^3.14.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v -s --ignore=framework"

[tool.poetry.scripts]
